+++ utilities/carp_study_generator/lib/cmd/command.dart
@@ -9,19 +9,20 @@ abstract class Command {
 /// An abstract class for all CARP Commands.
 abstract class AbstractCommand implements Command {
   static var _yaml;
-  CarpApp _app;
+  CarpApp? _app;
 
-  String get uri => _yaml['server']['uri'];
+  String? get uri => _yaml['server']['uri'];
   String get clientId => _yaml['server']['client_id'].toString();
   String get clientSecret => _yaml['server']['client_secret'].toString();
   String get username => _yaml['server']['username'].toString();
   String get password => _yaml['server']['password'].toString();
 
-  String get studyId => _yaml['study']['study_id'];
+  String? get studyId => _yaml['study']['study_id'];
   String get studyDeploymentId => _yaml['study']['study_deployment_id'];
 
   String get protocolPath => _yaml['protocol']['path'].toString();
   String get consentPath => _yaml['consent']['path'].toString();
+  String get descriptionPath => _yaml['description']['path'].toString();
 
   String get localizationPath => _yaml['localization']['path'].toString();
   List<dynamic> get locales => _yaml['localization']['locales'];
@@ -55,11 +56,11 @@ abstract class AbstractCommand implements Command {
   CarpApp get app {
     if (_app == null) {
       if (uri == null) throw Exception("No URI is provided");
-      if (!Uri.parse(uri).isAbsolute)
+      if (!Uri.parse(uri!).isAbsolute)
         throw Exception("Not a valid URI - '$uri'");
       try {
         if (studyId == null) throw Exception("A study ID must be provided");
-        if (studyId.length == 0)
+        if (studyId!.length == 0)
           throw Exception("The study ID cannot be empty - '$studyId'");
       } catch (e) {
         throw Exception("A valid study ID is not provided");
@@ -67,13 +68,13 @@ abstract class AbstractCommand implements Command {
 
       _app = CarpApp(
         name: "CARP server at '$uri'",
-        uri: Uri.parse(uri),
+        uri: Uri.parse(uri!),
         oauth: OAuthEndPoint(clientID: clientId, clientSecret: clientSecret),
         studyId: studyId,
         studyDeploymentId: studyDeploymentId,
       );
     }
-    return _app;
+    return _app!;
   }
 
   /// Authenticate at the CARP server.